+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+### Changed
+
+- Enable force upgrades when chart CR annotation is present.
+
 ## [1.0.4] - 2020-07-08
 
 ### Changed